+++ examples/gallery/overplot_hek_polygon.py
@@ -70,7 +70,8 @@ for i in range(0, n):
     rotated_boundary[i, 1] = new_coords[1].value
 
 ##############################################################################
-# Now let's plot the rotated coronal hole boundary on the AIA map.
+# Now let's plot the rotated coronal hole boundary on the AIA map, and fill
+# it with some matplotlib hatching.
 fig = plt.figure()
 ax = plt.subplot()
 aia_map.plot()