+++ sunpy/map/mapbase.py
@@ -4,46 +4,40 @@ Map is a generic Map class from which all other Map classes inherit from.
 from __future__ import absolute_import, division, print_function
 from sunpy.extern.six.moves import range
 
-#pylint: disable=E1101,E1121,W0404,W0613
-__authors__ = ["Russell Hewett, Stuart Mumford, Keith Hughitt, Steven Christe", "Jack Ireland"]
-__email__ = "stuart@mumford.me.uk"
-
 import warnings
 import inspect
 from abc import ABCMeta
 from copy import deepcopy
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import patches, cm, colors
 
 import astropy.wcs
+import astropy.units as u
 from astropy.coordinates import Longitude, Latitude
-
-from sunpy.image.transform import affine_transform
-from .nddata_compat import NDDataCompat as NDData
 
 import sunpy.io as io
 import sunpy.wcs as wcs
 import sunpy.coordinates
+from sunpy import config
+from sunpy.extern import six
 from sunpy.visualization import toggle_pylab, wcsaxes_compat
 from sunpy.sun import constants
 from sunpy.sun import sun
 from sunpy.time import parse_time, is_time
+from sunpy.image.transform import affine_transform
 from sunpy.image.rescale import reshape_image_to_4d_superpixel
 from sunpy.image.rescale import resample as sunpy_image_resample
 
-from sunpy.extern import six
+from .nddata_compat import NDDataCompat as NDData
 
-import astropy.units as u
-
-from collections import namedtuple
+TIME_FORMAT = config.get("general", "time_format")
 Pair = namedtuple('Pair', 'x y')
 
-
-from sunpy import config
-TIME_FORMAT = config.get("general", "time_format")
+__authors__ = "Russell Hewett, Stuart Mumford, Keith Hughitt," \
+              " Steven Christe, Jack Ireland"
 
 __all__ = ['GenericMap']
 
@@ -123,7 +117,7 @@ class GenericMap(NDData):
            [ 0.5   , -0.125 ,  0.4375, ...,  0.6875,  0.6875,  0.6875]])
 
 
-    >>> aia.units
+    >>> aia.spatial_units
     Pair(x=Unit("arcsec"), y=Unit("arcsec"))
     >>> aia.peek()   # doctest: +SKIP
 
@@ -138,10 +132,10 @@ class GenericMap(NDData):
     -----
 
     A number of the properties of this class are returned as two-value named
-    tuples that can either be indexed by position ([0] or [1]) or be accessed by
-    name (.x or .y).  The names "x" and "y" here refer to the first and second
-    axes of the map, and may not necessarily correspond to any similarly named
-    axes in the coordinate system.
+    tuples that can either be indexed by position ([0] or [1]) or be accessed
+    by name (.x or .y). The names "x" and "y" here refer to the first and
+    second axes of the map, and may not necessarily correspond to any similarly
+    named axes in the coordinate system.
 
     This class makes some assumptions about the WCS information contained in
     the meta data. The first and most extensive assumption is that it is
@@ -154,12 +148,14 @@ class GenericMap(NDData):
       notation using equations 32 in Thompson (2006).
 
     * If a CDi_j matrix is provided it is assumed that it can be converted to a
-      PCi_j matrix and CDELT keywords as described in Greisen & Calabretta (2002).
+      PCi_j matrix and CDELT keywords as described in
+      `Greisen & Calabretta (2002) <http://dx.doi.org/10.1051/0004-6361:20021327>`_
 
     * The 'standard' FITS keywords that are used by this class are the PCi_j
-      matrix and CDELT, along with the other keywords specified in the WCS papers.
-      All subclasses of this class must convert their header information to
-      this formalism. The CROTA to PCi_j conversion is done in this class.
+      matrix and CDELT, along with the other keywords specified in the WCS
+      papers. All subclasses of this class must convert their header
+      information to this formalism. The CROTA to PCi_j conversion is done in
+      this class.
 
     .. warning::
         This class currently assumes that a header with the CDi_j matrix
@@ -178,7 +174,7 @@ class GenericMap(NDData):
         self._fix_naxis()
 
         # Setup some attributes
-        self._nickname = self.detector
+        self._nickname = None
 
         # Validate header
         # TODO: This should be a function of the header, not of the map
@@ -199,7 +195,8 @@ class GenericMap(NDData):
     def __getitem__(self, key):
         """ This should allow indexing by physical coordinate """
         raise NotImplementedError(
-            "The ability to index Map by physical coordinate is not yet implemented.")
+            "The ability to index Map by physical"
+            " coordinate is not yet implemented.")
 
     def __repr__(self):
         if not self.observatory:
@@ -219,11 +216,11 @@ scale:\t\t {scale}
 
 """.format(dtype=self.__class__.__name__,
            obs=self.observatory, inst=self.instrument, det=self.detector,
-           meas=self.measurement, wave=self.wavelength, date=self.date, dt=self.exposure_time,
+           meas=self.measurement, wave=self.wavelength, date=self.date,
+           dt=self.exposure_time,
            dim=u.Quantity(self.dimensions),
            scale=u.Quantity(self.scale),
-           tmf=TIME_FORMAT)
-+ self.data.__repr__())
+           tmf=TIME_FORMAT) + self.data.__repr__())
 
     @property
     def wcs(self):
@@ -349,20 +346,30 @@ scale:\t\t {scale}
 
 # #### Keyword attribute and other attribute definitions #### #
 
-    @property
-    def name(self):
-        """Human-readable description of map-type"""
-        return "{obs} {detector} {measurement} {date:{tmf}}".format(obs=self.observatory,
-                                                                detector=self.detector,
-                                                                measurement=self.measurement,
+    def _base_name(self):
+        """Abstract the shared bit between name and latex_name"""
+        return "{nickname} {{measurement}} {date:{tmf}}".format(nickname=self.nickname,
                                                                 date=parse_time(self.date),
                                                                 tmf=TIME_FORMAT)
+
+    @property
+    def name(self):
+        """Human-readable description of the Map."""
+        return self._base_name().format(measurement=self.measurement)
+
+    @property
+    def latex_name(self):
+        """LaTeX formatted description of the Map."""
+        if isinstance(self.measurement, u.Quantity):
+            return self._base_name().format(measurement=self.measurement._repr_latex_())
+        else:
+            return self.name
 
     @property
     def nickname(self):
         """An abbreviated human-readable description of the map-type; part of
         the Helioviewer data model"""
-        return self._nickname
+        return self._nickname if self._nickname else self.detector
 
     @nickname.setter
     def nickname(self, n):
@@ -373,16 +380,11 @@ scale:\t\t {scale}
         """Image observation time"""
         time = parse_time(self.meta.get('date-obs', 'now'))
         if time is None:
-            warnings.warn_explicit("Missing metadata for observation time. Using current time.",
-                                       Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for observation time."
+                                   " Using current time.",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
         return parse_time(time)
-
-#    @date.setter
-#    def date(self, new_date):
-#        self.meta['date-obs'] = new_date
-#        #propagate change to malformed FITS keywords
-#        if is_time(self.meta.get('date_obs', None)):
-#            self.meta['date_obs'] = new_date
 
     @property
     def detector(self):
@@ -395,8 +397,10 @@ scale:\t\t {scale}
         dsun = self.meta.get('dsun_obs', None)
 
         if dsun is None:
-            warnings.warn_explicit("Missing metadata for Sun-spacecraft separation: assuming Sun-Earth distance",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for Sun-spacecraft"
+                                   " separation: assuming Sun-Earth distance",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             dsun = sun.sunearth_distance(self.date).to(u.m)
 
         return u.Quantity(dsun, 'm')
@@ -414,22 +418,25 @@ scale:\t\t {scale}
     @property
     def measurement(self):
         """Measurement name, defaults to the wavelength of image"""
-        return u.Quantity(self.meta.get('wavelnth', 0), self.meta.get('waveunit', ""))
+        return u.Quantity(self.meta.get('wavelnth', 0),
+                          self.meta.get('waveunit', ""))
 
     @property
     def wavelength(self):
         """wavelength of the observation"""
-        return u.Quantity(self.meta.get('wavelnth', 0), self.meta.get('waveunit', ""))
+        return u.Quantity(self.meta.get('wavelnth', 0),
+                          self.meta.get('waveunit', ""))
 
     @property
     def observatory(self):
         """Observatory or Telescope name"""
-        return self.meta.get('obsrvtry', self.meta.get('telescop', "")).replace("_", " ")
+        return self.meta.get('obsrvtry',
+                             self.meta.get('telescop', "")).replace("_", " ")
 
     @property
     def xrange(self):
         """Return the X range of the image from edge to edge."""
-        #TODO: This should be reading from the WCS object
+        # TODO: This should be reading from the WCS object
         xmin = self.center.x - self.dimensions[0] / 2. * self.scale.x
         xmax = self.center.x + self.dimensions[0] / 2. * self.scale.x
         return u.Quantity([xmin, xmax])
@@ -437,14 +444,16 @@ scale:\t\t {scale}
     @property
     def yrange(self):
         """Return the Y range of the image from edge to edge."""
-        #TODO: This should be reading from the WCS object
+        # TODO: This should be reading from the WCS object
         ymin = self.center.y - self.dimensions[1] / 2. * self.scale.y
         ymax = self.center.y + self.dimensions[1] / 2. * self.scale.y
         return u.Quantity([ymin, ymax])
 
     @property
     def center(self):
-        """The offset between the center of the Sun and the center of the map."""
+        """
+        The offset between the center of the Sun and the center of the map.
+        """
         return Pair(wcs.get_center(self.dimensions[0], self.scale.x,
                                    self.reference_pixel.x,
                                    self.reference_coordinate.x),
@@ -460,9 +469,11 @@ scale:\t\t {scale}
 
     @u.quantity_input(x=u.deg, y=u.deg)
     def shift(self, x, y):
-        """Returns a map shifted by a specified amount to, for example, correct for a bad
-        map location. These values are applied directly to the `~sunpy.map.GenericMap.reference_coordinate`.
-        To check how much shift has already been applied see `~sunpy.map.GenericMap.shifted_value`
+        """
+        Returns a map shifted by a specified amount to, for example, correct
+        for a bad map location. These values are applied directly to the
+        `~sunpy.map.GenericMap.reference_coordinate`. To check how much shift
+        has already been applied see `~sunpy.map.GenericMap.shifted_value`
 
         Parameters
         ----------
@@ -479,13 +490,16 @@ scale:\t\t {scale}
             A new shifted Map.
         """
         new_map = deepcopy(self)
-        new_map._shift = Pair(self.shifted_value.x + x, self.shifted_value.y + y)
+        new_map._shift = Pair(self.shifted_value.x + x,
+                              self.shifted_value.y + y)
 
         new_meta = self.meta.copy()
 
         # Update crvals
-        new_meta['crval1'] = ((self.meta['crval1'] * self.spatial_units.x + x).to(self.spatial_units.x)).value
-        new_meta['crval2'] = ((self.meta['crval2'] * self.spatial_units.y + y).to(self.spatial_units.y)).value
+        new_meta['crval1'] = ((self.meta['crval1'] *
+                               self.spatial_units.x + x).to(self.spatial_units.x)).value
+        new_meta['crval2'] = ((self.meta['crval2'] *
+                               self.spatial_units.y + y).to(self.spatial_units.y)).value
 
         new_map.meta = new_meta
 
@@ -501,11 +515,14 @@ scale:\t\t {scale}
         """Radius of the Sun."""
         rsun_arcseconds = self.meta.get('rsun_obs',
                                         self.meta.get('solar_r',
-                                                      self.meta.get('radius', None)))
+                                                      self.meta.get('radius',
+                                                                    None)))
 
         if rsun_arcseconds is None:
-            warnings.warn_explicit("Missing metadata for solar radius: assuming photospheric limb as seen from Earth",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for solar radius:"
+                                   " assuming photospheric limb as seen from Earth",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             rsun_arcseconds = sun.solar_semidiameter_angular_size(self.date).to('arcsec').value
 
         return u.Quantity(rsun_arcseconds, 'arcsec')
@@ -522,8 +539,10 @@ scale:\t\t {scale}
         carrington_longitude = self.meta.get('crln_obs', None)
 
         if carrington_longitude is None:
-            warnings.warn_explicit("Missing metadata for Carrington longitude: assuming Earth-based observer",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for Carrington longitude:"
+                                   " assuming Earth-based observer",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             carrington_longitude = (sun.heliographic_solar_center(self.date))[0]
 
         return u.Quantity(carrington_longitude, 'deg')
@@ -536,8 +555,10 @@ scale:\t\t {scale}
                                                             self.meta.get('solar_b0', None)))
 
         if heliographic_latitude is None:
-            warnings.warn_explicit("Missing metadata for heliographic latitude: assuming Earth-based observer",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for heliographic latitude:"
+                                   " assuming Earth-based observer",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             heliographic_latitude = (sun.heliographic_solar_center(self.date))[1]
 
         return u.Quantity(heliographic_latitude, 'deg')
@@ -557,26 +578,34 @@ scale:\t\t {scale}
     @property
     def reference_pixel(self):
         """Reference point axes in pixels (i.e. crpix1, crpix2)"""
-        return Pair(self.meta.get('crpix1', (self.meta.get('naxis1') + 1) / 2.) * u.pixel,
-                    self.meta.get('crpix2', (self.meta.get('naxis2') + 1) / 2.) * u.pixel)
+        return Pair(self.meta.get('crpix1',
+                                  (self.meta.get('naxis1') + 1) / 2.) * u.pixel,
+                    self.meta.get('crpix2',
+                                  (self.meta.get('naxis2') + 1) / 2.) * u.pixel)
 
     @property
     def scale(self):
-        """Image scale along the x and y axes in units/pixel (i.e. cdelt1, cdelt2)"""
-        #TODO: Fix this if only CDi_j matrix is provided
+        """
+        Image scale along the x and y axes in units/pixel (i.e. cdelt1, cdelt2)
+        """
+        # TODO: Fix this if only CDi_j matrix is provided
         return Pair(self.meta.get('cdelt1', 1.) * self.spatial_units.x / u.pixel,
                     self.meta.get('cdelt2', 1.) * self.spatial_units.y / u.pixel)
 
     @property
     def spatial_units(self):
-        """Image coordinate units along the x and y axes (i.e. cunit1, cunit2)."""
+        """
+        Image coordinate units along the x and y axes (i.e. cunit1, cunit2).
+        """
         return Pair(u.Unit(self.meta.get('cunit1', 'arcsec')),
                     u.Unit(self.meta.get('cunit2', 'arcsec')))
 
     @property
     def rotation_matrix(self):
-        """Matrix describing the rotation required to align solar North with
-        the top of the image."""
+        """
+        Matrix describing the rotation required to align solar North with
+        the top of the image.
+        """
         if 'PC1_1' in self.meta:
             return np.matrix([[self.meta['PC1_1'], self.meta['PC1_2']],
                               [self.meta['PC2_1'], self.meta['PC2_2']]])
@@ -608,8 +637,8 @@ scale:\t\t {scale}
 # #### Miscellaneous #### #
 
     def _fix_date(self):
-        # Check commonly used but non-standard FITS keyword for observation time
-        # and correct the keyword if we can.  Keep updating old one for
+        # Check commonly used but non-standard FITS keyword for observation
+        # time and correct the keyword if we can. Keep updating old one for
         # backwards compatibility.
         if is_time(self.meta.get('date_obs', None)):
             self.meta['date-obs'] = self.meta['date_obs']
@@ -638,7 +667,8 @@ scale:\t\t {scale}
 
     def _get_cmap_name(self):
         """Build the default color map name."""
-        cmap_string = self.observatory + self.meta['detector'] + str(int(self.wavelength.to('angstrom').value))
+        cmap_string = (self.observatory + self.meta['detector'] +
+                       str(int(self.wavelength.to('angstrom').value)))
         return cmap_string.lower()
 
     def _validate_meta(self):
@@ -655,14 +685,13 @@ scale:\t\t {scale}
 
         """
 
-        warnings.simplefilter('always', Warning)
-
         for meta_property in ('cunit1', 'cunit2', 'waveunit'):
             if (self.meta.get(meta_property) and
                 u.Unit(self.meta.get(meta_property),
                        parse_strict='silent').physical_type == 'unknown'):
 
-                warnings.warn("Unknown value for "+meta_property.upper(), Warning)
+                warnings.warn("Unknown value for " + meta_property.upper(),
+                              Warning)
 
 
 # #### Data conversion routines #### #
@@ -696,7 +725,8 @@ scale:\t\t {scale}
         y : `~astropy.units.Quantity`
             Pixel coordinate on the CTYPE2 axis.
         """
-        x, y = self.wcs.wcs_world2pix(x.to(u.deg).value, y.to(u.deg).value, origin)
+        x, y = self.wcs.wcs_world2pix(x.to(u.deg).value,
+                                      y.to(u.deg).value, origin)
 
         return x * u.pixel, y * u.pixel
 
@@ -797,7 +827,8 @@ scale:\t\t {scale}
 
         References
         ----------
-        * `Rebinning <http://www.scipy.org/Cookbook/Rebinning>`_ (Original source, 2011/11/19)
+        * `Rebinning <http://www.scipy.org/Cookbook/Rebinning>`_ (Original
+          source, 2011/11/19)
         """
 
         # Note: because the underlying ndarray is transposed in sense when
@@ -808,7 +839,7 @@ scale:\t\t {scale}
 
         # Make a copy of the original data and perform resample
         new_data = sunpy_image_resample(self.data.copy().T, dimensions,
-                                    method, center=True)
+                                        method, center=True)
         new_data = new_data.T
 
         scale_factor_x = float(self.dimensions[0] / dimensions[0])
@@ -857,13 +888,13 @@ scale:\t\t {scale}
         rmatrix : 2x2
             Linear transformation rotation matrix.
         order : int 0-5
-            Interpolation order to be used. When using scikit-image this parameter
-            is passed into :func:`skimage.transform.warp` (e.g., 4 corresponds to
-            bi-quartic interpolation).
+            Interpolation order to be used. When using scikit-image this
+            parameter is passed into :func:`skimage.transform.warp` (e.g., 4
+            corresponds to bi-quartic interpolation).
             When using scipy it is passed into
-            :func:`scipy.ndimage.interpolation.affine_transform` where it controls
-            the order of the spline.
-            Faster performance may be obtained at the cost of accuracy by using lower values.
+            :func:`scipy.ndimage.interpolation.affine_transform` where it
+            controls the order of the spline. Faster performance may be
+            obtained at the cost of accuracy by using lower values.
             Default: 4
         scale : float
             A scale factor for the image, default is no scaling
@@ -887,7 +918,8 @@ scale:\t\t {scale}
 
         See Also
         --------
-        sunpy.image.transform.affine_transform : The routine this method calls for the rotation.
+        sunpy.image.transform.affine_transform : The routine this method calls
+        for the rotation.
 
         Notes
         -----
@@ -895,8 +927,8 @@ scale:\t\t {scale}
         This function will also convert a CDi_j matrix to a PCi_j matrix.
 
         See :func:`sunpy.image.transform.affine_transform` for details on the
-        transformations, situations when the underlying data is modified prior to rotation,
-        and differences from IDL's rot().
+        transformations, situations when the underlying data is modified prior
+        to rotation, and differences from IDL's rot().
         """
         if angle is not None and rmatrix is not None:
             raise ValueError("You cannot specify both an angle and a matrix")
@@ -913,7 +945,7 @@ scale:\t\t {scale}
                     raise u.UnitsError("Argument '{0}' to function '{1}'"
                                        " must be in units convertable to"
                                        " '{2}'.".format('angle', 'rotate',
-                                                      u.deg.to_string()))
+                                                        u.deg.to_string()))
 
             # Either there is no .unit or no .is_equivalent
             except AttributeError:
@@ -922,8 +954,8 @@ scale:\t\t {scale}
                 else:
                     error_msg = "no 'unit' attribute"
                 raise TypeError("Argument '{0}' to function '{1}' has {2}. "
-                      "You may want to pass in an astropy Quantity instead."
-                         .format('angle', 'rotate', error_msg))
+                                "You may want to pass in an astropy Quantity instead."
+                                 .format('angle', 'rotate', error_msg))
 
         # Interpolation parameter sanity
         if order not in range(6):
@@ -947,7 +979,7 @@ scale:\t\t {scale}
         extent = np.max(np.abs(np.vstack((new_map.data.shape * rmatrix,
                                           new_map.data.shape * rmatrix.T))), axis=0)
         # Calculate the needed padding or unpadding
-        diff = np.asarray(np.ceil((extent - new_map.data.shape) / 2)).ravel()
+        diff = np.asarray(np.ceil((extent - new_map.data.shape) / 2), dtype=int).ravel()
         # Pad the image array
         pad_x = int(np.max((diff[1], 0)))
         pad_y = int(np.max((diff[0], 0)))
@@ -1594,7 +1626,7 @@ scale:\t\t {scale}
         if 'title' in imshow_args:
             plot_settings_title = imshow_args.pop('title')
         else:
-            plot_settings_title = self.name
+            plot_settings_title = self.latex_name
 
         if annotate:
             if title is True: